Adds Redis test for basic functionality, game logic start. Remove setup.js from testing
