+++ package.json
@@ -62,6 +62,9 @@
   "jest": {
     "setupFiles": [
       "<rootDir>__tests__/setup.js"
+    ],
+    "testPathIgnorePatterns": [
+      "<rootDir>/__tests__/setup.js"
     ]
   }
 }
